requirements.txt
Flask==2.1.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
flask-restx==1.1.0
Pyrebase4==4.7.1
python-dotenv==1.0.0
Werkzeug==2.2.2
yake==0.4.8
nltk==3.8.1
wikipedia-api==0.6.0